+++ Reach/src/utopia/reach/component/template/CursorDefining.scala
@@ -15,13 +15,18 @@ object CursorDefining
 	  * @param component Component to define
 	  * @param cursorTypePointer Pointer to the type of cursor to use
 	  * @param shadePointer Pointer to the shade to expect the component to be (light or dark)
+	 *  @param customization A function that accepts a cursor and the cursor position
+	 *                       (within and relative to component's bounds) and returns a custom cursor image to apply,
+	 *                       if appropriate. Returns None if no custom cursor should be applied.
+	 *                       None if no customization function should be used.
 	  */
 	def defineCursorFor(component: ReachComponentLike, cursorTypePointer: View[CursorType],
-	                    shadePointer: View[ColorShadeVariant]) =
+	                    shadePointer: View[ColorShadeVariant],
+	                    customization: Option[(Cursor, Point) => Option[Image]] = None) =
 	{
 		// Only works if cursor management is enabled in canvas
 		component.parentCanvas.cursorManager.foreach { manager =>
-			val wrapped = new CursorDefiningWrapper(component, cursorTypePointer, shadePointer)
+			val wrapped = new CursorDefiningWrapper(component, cursorTypePointer, shadePointer, customization)
 			// Adds the component to the manager whenever it is attached to the main stack hierarchy
 			component.addHierarchyListener { isAttached =>
 				if (isAttached)
@@ -31,7 +36,6 @@ object CursorDefining
 			}
 		}
 	}
-	
 	/**
 	  * Specifies cursor settings for a component that doesn't normally define cursor
 	  * @param component Component to define
@@ -39,19 +43,31 @@ object CursorDefining
 	  * @param shade shade to expect the component to be (light or dark)
 	  */
 	def defineCursorFor(component: ReachComponentLike, cursorType: CursorType, shade: ColorShadeVariant): Unit =
-		defineCursorFor(component, View(cursorType), View(shade))
+		defineCursorFor(component, View.fixed(cursorType), View.fixed(shade))
 	
 	
 	// NESTED	--------------------------------
 	
 	private class CursorDefiningWrapper(wrapped: ReachComponentLike, cursorTypePointer: View[CursorType],
-	                                    shadePointer: View[ColorShadeVariant]) extends CursorDefining
+	                                    shadePointer: View[ColorShadeVariant],
+	                                    customization: Option[(Cursor, Point) => Option[Image]] = None)
+		extends CursorDefining
 	{
+		// IMPLEMENTED  -------------------------
+		
 		override def cursorType = cursorTypePointer.value
 		
 		override def cursorBounds = wrapped.boundsInsideTop
 		
-		override def cursorToImage(cursor: Cursor, position: Point) = cursor.over(shadePointer.value)
+		override def cursorToImage(cursor: Cursor, position: Point) = customization match {
+			case Some(custom) => custom(cursor, position).getOrElse { defaultCursorToImage(cursor) }
+			case None => defaultCursorToImage(cursor)
+		}
+		
+		
+		// OTHER    ----------------------------
+		
+		private def defaultCursorToImage(cursor: Cursor) = cursor.over(shadePointer.value)
 	}
 	
 	
@@ -63,8 +79,7 @@ object CursorDefining
 		 * Registers this component to the parent canvases cursor manager, if available.
 		 * Only keeps this component managed while attached to the main stack hierarchy.
 		 */
-		def register() =
-		{
+		def register() = {
 			c.parentCanvas.cursorManager.foreach { manager =>
 				c.addHierarchyListener { isAttached => if (isAttached) manager += c else manager -= c }
 			}