+++ Paradigm/Changes.md
@@ -3,6 +3,9 @@
 ## v1.2 (In Development)
 This update reflects in style the **Flow** v2.0 update. Multiple classes and concepts have been rewritten 
 and a more intuitive naming logic is applied.
+### New Methods
+- **Bounds** (type)
+  - Added `.between(Pair)`
 
 While there are a fair number of breaking changes, you shouldn't encounter too many, unless you have created 
 your own shape classes which extend from the base traits.